Extract command implementations into XcodesKit
The motivation for this is to be able to test more of the command implementations, and particularly logging during installation.

+++ Sources/XcodesKit/XcodeInstaller.swift
@@ -2,6 +2,8 @@ import Foundation
 import PromiseKit
 import Path
 import AppleAPI
+import Version
+import LegibleError
 
 /// Downloads and installs Xcodes
 public final class XcodeInstaller {
@@ -14,6 +16,12 @@ public final class XcodeInstaller {
         case codesignVerifyFailed(output: String)
         case unexpectedCodeSigningIdentity(identifier: String, certificateAuthority: [String])
         case unsupportedFileFormat(extension: String)
+        case missingSudoerPassword
+        case unavailableVersion(Version)
+        case missingUsernameOrPassword
+        case versionAlreadyInstalled(InstalledXcode)
+        case invalidVersion(String)
+        case versionNotInstalled(Version)
 
         public var errorDescription: String? {
             switch self {
@@ -42,11 +50,186 @@ public final class XcodeInstaller {
                        """
             case .unsupportedFileFormat(let fileExtension):
                 return "xcodes doesn't (yet) support installing Xcode from the \(fileExtension) file format."
+            case .missingSudoerPassword:
+                return "Missing password. Please try again."
+            case let .unavailableVersion(version):
+                return "Could not find version \(version.xcodeDescription)."
+            case .missingUsernameOrPassword:
+                return "Missing username or a password. Please try again."
+            case let .versionAlreadyInstalled(installedXcode):
+                return "\(installedXcode.version.xcodeDescription) is already installed at \(installedXcode.path)"
+            case let .invalidVersion(version):
+                return "\(version) is not a valid version number."
+            case let .versionNotInstalled(version):
+                return "\(version.xcodeDescription) is not installed."
             }
         }
     }
 
-    public init() {}
+    private var configuration: Configuration
+    private var xcodeList: XcodeList
+
+    public init(configuration: Configuration, xcodeList: XcodeList) {
+        self.configuration = configuration
+        self.xcodeList = xcodeList
+    }
+
+    public func install(_ versionString: String, _ urlString: String?) -> Promise<Void> {
+        return firstly { () -> Promise<(Xcode, URL)> in
+            guard let version = Version(xcodeVersion: versionString) ?? versionFromXcodeVersionFile() else {
+                throw Error.invalidVersion(versionString)
+            }
+
+            if let installedXcode = Current.files.installedXcodes().first(where: { $0.version.isEqualWithoutBuildMetadataIdentifiers(to: version) }) {
+                throw Error.versionAlreadyInstalled(installedXcode)
+            }
+
+            if let urlString = urlString {
+                let url = URL(fileURLWithPath: urlString, relativeTo: nil)
+                let xcode = Xcode(version: version, url: url, filename: String(url.path.suffix(fromLast: "/")))
+                return Promise.value((xcode, url))
+            }
+            else {
+                return self.downloadXcode(version: version)
+            }
+        }
+        .then { xcode, url -> Promise<Void> in
+            return self.installArchivedXcode(xcode, at: url, archiveTrashed: { archiveURL in
+                Current.logging.log("Xcode archive \(url.lastPathComponent) has been moved to the Trash.")
+            }, passwordInput: { () -> Promise<String> in
+                return Promise { seal in
+                    Current.logging.log("xcodes requires superuser privileges in order to setup some parts of Xcode.")
+                    guard let password = Current.shell.readSecureLine(prompt: "Password: ") else { seal.reject(Error.missingSudoerPassword); return }
+                    seal.fulfill(password + "\n")
+                }
+            })
+        }
+        .done {
+            Current.shell.exit(0)
+        }
+    }
+
+    private func versionFromXcodeVersionFile() -> Version? {
+        let xcodeVersionFilePath = Path.cwd.join(".xcode-version")
+        let version = (try? Data(contentsOf: xcodeVersionFilePath.url))
+            .flatMap { String(data: $0, encoding: .utf8) }
+            .flatMap(Version.init(gemVersion:))
+        return version
+    }
+
+    private func downloadXcode(version: Version) -> Promise<(Xcode, URL)> {
+        return firstly { () -> Promise<Version> in
+            loginIfNeeded().map { version }
+        }
+        .then { version -> Promise<Version> in
+            if self.xcodeList.shouldUpdate {
+                return self.xcodeList.update().map { _ in version }
+            }
+            else {
+                return Promise.value(version)
+            }
+        }
+        .then { version -> Promise<(Xcode, URL)> in
+            guard let xcode = self.xcodeList.availableXcodes.first(where: { version.isEqualWithoutBuildMetadataIdentifiers(to: $0.version) }) else {
+                throw Error.unavailableVersion(version)
+            }
+
+            // Move to the next line
+            Current.logging.log("")
+            let formatter = NumberFormatter(numberStyle: .percent)
+            var observation: NSKeyValueObservation?
+
+            let promise = self.downloadOrUseExistingArchive(for: xcode, progressChanged: { progress in
+                observation?.invalidate()
+                observation = progress.observe(\.fractionCompleted) { progress, _ in
+                    // These escape codes move up a line and then clear to the end
+                    Current.logging.log("\u{1B}[1A\u{1B}[K" + "Downloading Xcode \(xcode.version): " + formatter.string(from: progress.fractionCompleted)!)
+                }
+            })
+
+            return promise
+                .get { _ in observation?.invalidate() }
+                .map { return (xcode, $0) }
+        }
+    }
+
+    func loginIfNeeded(withUsername existingUsername: String? = nil) -> Promise<Void> {
+        return firstly { () -> Promise<Void> in
+            return Current.network.validateSession()
+        }
+        .recover { error -> Promise<Void> in
+            guard
+                let username = existingUsername ?? self.findUsername() ?? Current.shell.readLine(prompt: "Apple ID: "),
+                let password = self.findPassword(withUsername: username) ?? Current.shell.readSecureLine(prompt: "Apple ID Password: ")
+            else { throw Error.missingUsernameOrPassword }
+
+            return firstly { () -> Promise<Void> in
+                self.login(username, password: password)
+            }
+            .recover { error -> Promise<Void> in
+                Current.logging.log(error.legibleLocalizedDescription)
+
+                if case Client.Error.invalidUsernameOrPassword = error {
+                    Current.logging.log("Try entering your password again")
+                    return self.loginIfNeeded(withUsername: username)
+                }
+                else {
+                    return Promise(error: error)
+                }
+            }
+        }
+    }
+
+    func login(_ username: String, password: String) -> Promise<Void> {
+        return firstly { () -> Promise<Void> in
+            Current.network.login(accountName: username, password: password)
+        }
+        .recover { error -> Promise<Void> in
+
+            if let error = error as? Client.Error {
+              switch error  {
+              case .invalidUsernameOrPassword(_):
+                  // remove any keychain password if we fail to log with an invalid username or password so it doesn't try again.
+                  try? Current.keychain.remove(username)
+              default:
+                  break
+              }
+            }
+
+            return Promise(error: error)
+        }
+        .done { _ in
+            try? Current.keychain.set(password, key: username)
+
+            if self.configuration.defaultUsername != username {
+                self.configuration.defaultUsername = username
+                try? self.configuration.save()
+            }
+        }
+    }
+
+    let xcodesUsername = "XCODES_USERNAME"
+    let xcodesPassword = "XCODES_PASSWORD"
+
+    func findUsername() -> String? {
+        if let username = Current.shell.env(xcodesUsername) {
+            return username
+        }
+        else if let username = configuration.defaultUsername {
+            return username
+        }
+        return nil
+    }
+
+    func findPassword(withUsername username: String) -> String? {
+        if let password = Current.shell.env(xcodesPassword) {
+            return password
+        }
+        else if let password = try? Current.keychain.getString(username){
+            return password
+        }
+        return nil
+    }
 
     public func downloadOrUseExistingArchive(for xcode: Xcode, progressChanged: @escaping (Progress) -> Void) -> Promise<URL> {
         // Check to see if the archive is in the expected path in case it was downloaded but failed to install
@@ -115,10 +298,67 @@ public final class XcodeInstaller {
         }
     }
 
-    public func uninstallXcode(_ installedXcode: InstalledXcode) -> Promise<URL> {
-        return Promise<URL> { seal in
-            seal.fulfill(try Current.files.trashItem(at: installedXcode.path.url))
+    public func uninstallXcode(_ versionString: String) -> Promise<Void> {
+        return firstly { () -> Promise<(InstalledXcode, URL)> in
+            guard let version = Version(xcodeVersion: versionString) else {
+                throw Error.invalidVersion(versionString)
+            }
+
+            guard let installedXcode = Current.files.installedXcodes().first(where: { $0.version.isEqualWithoutBuildMetadataIdentifiers(to: version) }) else {
+                throw Error.versionNotInstalled(version)
+            }
+
+            return Promise<URL> { seal in
+                seal.fulfill(try Current.files.trashItem(at: installedXcode.path.url))
+            }.map { (installedXcode, $0) }
         }
+        .done { (installedXcode, trashURL) in
+            Current.logging.log("Xcode \(installedXcode.version.xcodeDescription) moved to Trash: \(trashURL.path)")
+            Current.shell.exit(0)
+        }
+    }
+
+    public func updateAndPrint() -> Promise<Void> {
+        return firstly { () -> Promise<Void> in
+            loginIfNeeded()
+        }
+        .then { () -> Promise<[Xcode]> in
+            self.xcodeList.update()
+        }
+        .done { xcodes in
+            self.printAvailableXcodes(xcodes, installed: Current.files.installedXcodes())
+            Current.shell.exit(0)
+        }
+    }
+
+    public func printAvailableXcodes(_ xcodes: [Xcode], installed installedXcodes: [InstalledXcode]) {
+        var allXcodeVersions = xcodes.map { $0.version }
+        for installedXcode in installedXcodes {
+            // If an installed version isn't listed online, add the installed version
+            if !allXcodeVersions.contains(where: { version in
+                version.isEquivalentForDeterminingIfInstalled(toInstalled: installedXcode.version)
+            }) {
+                allXcodeVersions.append(installedXcode.version)
+            }
+            // If an installed version is the same as one that's listed online which doesn't have build metadata, replace it with the installed version with build metadata
+            else if let index = allXcodeVersions.firstIndex(where: { version in
+                version.isEquivalentForDeterminingIfInstalled(toInstalled: installedXcode.version) &&
+                version.buildMetadataIdentifiers.isEmpty
+            }) {
+                allXcodeVersions[index] = installedXcode.version
+            }
+        }
+
+        allXcodeVersions
+            .sorted { $0 < $1 }
+            .forEach { xcodeVersion in
+                if installedXcodes.contains(where: { xcodeVersion.isEquivalentForDeterminingIfInstalled(toInstalled: $0.version) }) {
+                    Current.logging.log("\(xcodeVersion.xcodeDescription) (Installed)")
+                }
+                else {
+                    Current.logging.log(xcodeVersion.xcodeDescription)
+                }
+            }
     }
 
     func unarchiveAndMoveXIP(at source: URL, to destination: URL) -> Promise<URL> {
